tests/wpt/tests/css/css-position/inheritance.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>Inheritance of CSS Positioned Layout properties</title>
<link rel="help" href="https://drafts.csswg.org/css-position/#property-index">
<meta name="assert" content="Properties inherit or not according to the spec.">
<meta name="assert" content="Properties have initial values according to the spec.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/inheritance-testcommon.js"></script>
</head>
<body>
<div id="container">
<div id="target"></div>
</div>
<style>
#container, #target {
  position: sticky;
}
</style>
<script>
assert_not_inherited('position', 'static', 'absolute');
assert_not_inherited('top', 'auto', '10px');
assert_not_inherited('right', 'auto', '10px');
assert_not_inherited('bottom', 'auto', '10px');
assert_not_inherited('left', 'auto', '10px');
assert_not_inherited('z-index', 'auto', '123');
</script>
</body>
</html>
